+++ packages/devextreme/js/__internal/scheduler/appointments/data_provider/utils/get_appointments_occurrences.ts
@@ -3,7 +3,7 @@ import { dateUtilsTs } from '@ts/core/utils/date';
 import { getRecurrenceProcessor } from '../../../m_recurrence';
 import type { TimeZoneCalculator } from '../../../r1/timezone_calculator';
 import type { AppointmentDataItem } from '../../../types';
-import { getRecurrenceException } from './getRecurrenceException';
+import { getRecurrenceException } from './get_recurrence_exception';
 import type { DateInterval } from './type';
 
 interface Options {
