+++ packages/devextreme/js/__internal/scheduler/appointments/data_provider/utils/get_recurrence_exception.ts
