+++ packages/devextreme/js/__internal/scheduler/appointments/data_provider/utils/get_visible_date_time_intervals.ts
