+++ packages/devextreme/js/__internal/scheduler/appointments/data_provider/utils/index.ts
@@ -1,1 +1,1 @@
-export { getAppointmentFilter } from './getAppointmentFilter';
+export { getAppointmentFilter } from './get_appointment_filter';
