+++ packages/devextreme/js/__internal/scheduler/appointments/data_provider/utils/is_appointment_matched_intervals.test.ts
@@ -4,7 +4,7 @@ import {
 
 import {
   isAppointmentMatchedIntervals,
-} from './isAppointmentMatchedIntervals';
+} from './is_appointment_matched_intervals';
 
 const viewIntervals = [{
   min: new Date(2000, 0, 10),
