+++ packages/devextreme/js/__internal/scheduler/appointments/data_provider/utils/is_appointment_matched_intervals.ts
